rename Autotune to Benchmark. remove examples.
