+++ src/tune/metrics.py
@@ -1,0 +1,37 @@
+from typing import Tuple
+
+
+def calculate_pe_utilization(
+    lhsT_shape: Tuple[int, ...], rhs_shape: Tuple[int, ...], time_ms: float, target_instance_family: str
+) -> float:
+    """
+    Calculate hardware FLOPS utilization for a GEMM operation.
+
+    Parameters:
+    lhsT_shape (tuple): Shape of matrix A (k, m)
+    rhs_shape (tuple): Shape of matrix B (k, n)
+    time_ms (float): Execution time in milliseconds
+
+    Returns:
+    dict: Dictionary containing FLOPS, TFLOPS, max TFLOPS, and utilization percentage
+    """
+    k, m = lhsT_shape
+    _k, n = rhs_shape
+    assert k == _k, f"Incompatible matrix dimensions: {lhsT_shape} and {rhs_shape}"
+
+    if any(target_instance_family.startswith(family) for family in {"sunda", "trainium", "trn1", "inf2"}):
+        pe_freq = 2.8 * 1e9  # Hz (2.8 GHz)
+        num_lnc = 1
+    elif any(target_instance_family.startswith(family) for family in {"trn2", "inf3", "gen3", "cayman"}):
+        pe_freq = 2.4 * 1e9  # Hz (2.4 GHz)
+        num_lnc = 2
+    else:
+        raise NotImplementedError("Unknown target instance: " + target_instance_family)
+
+    # Calculate total FLOPS (2 operations per matrix element - multiply and add)
+    flops = 2 * m * k * n
+    actual_latency_s = time_ms / 1000
+    actual_pe_cycles = actual_latency_s * pe_freq
+    theoretical_pe_cycles = flops / (2 * 128 * 128 * num_lnc)
+    pe_utilization = theoretical_pe_cycles / actual_pe_cycles
+    return pe_utilization
